add missing status codes
